rename discovery -> discoveryButtons
