+++ src/components/DiscoveryButtons.vue
